migrate participation methods to IS class; FSS is still TODO;
